pyproject.toml
[project]
name = "memory-workspace"
version = "0.0.0"
description = "Memory SDK workspace"
requires-python = ">=3.12.4"

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
recollect = { workspace = true }
recollect-mcp = { workspace = true }

[dependency-groups]
dev = [
    "recollect",
    "recollect-mcp",
    "asyncpg-stubs>=0.31.2",
    "hypothesis>=6.151.9",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
    "rich>=14.3.3",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true
namespace_packages = true
show_error_codes = true
show_column_numbers = true


[tool.ruff]
target-version = "py312"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C90", # mccabe complexity
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "YTT", # flake8-2020
    "S",   # flake8-bandit
    "BLE", # flake8-blind-except
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
    "C4",  # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "G",   # flake8-logging-format
    "PIE", # flake8-pie
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "RET", # flake8-return
    "PD",  # pandas-vet
    "PGH", # pygrep-hooks
    "RUF", # ruff-specific
]
ignore = [
    "S101",  # assert used
    "G004",  # logging f-string
]

[tool.ruff.lint.per-file-ignores]
"examples/*" = ["T201", "BLE001", "S110"]

[tool.ruff.lint.mccabe]
max-complexity = 10  # Enforce low complexity

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers -m 'not slow'"
testpaths = ["tests"]
pythonpath = ["packages/memory/src", "packages/memory-mcp/src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore:The `__fields__` attribute is deprecated:pydantic.warnings.PydanticDeprecatedSince20",
    "ignore:The configuration option \"asyncio_default_fixture_loop_scope\" is unset:pytest.PytestDeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "agent: marks tests that require LLM agents (deselect with '-m \"not agent\"')"
]

[tool.coverage.run]
source = ["packages/memory/src/recollect", "packages/memory-mcp/src/recollect_mcp"]
omit = ["*/tests/*", "*/examples/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
